reorganize code, add logout
