+++ server/controllers/index.js
@@ -2,10 +2,12 @@ const todos = require('./todos');
 const todoItems = require('./todoitems');
 const projectsController = require('./projectsController');
 const usersController = require('./usersController');
+const authController = require('./authController');
 
 module.exports = {
   todos,
   todoItems,
   projectsController,
-  usersController
+  usersController,
+  authController
 };
